Enable actual thread-locals.
This switches to using #[thread_local] instead of just declaring static
mut variables and *assuming* they are only accessed from one thread.

The decision to have only a single thread is now made in main, which
defines the method for accessing thread-local storage, and provides the
linker script that generates a single copy of thread locals.