Changes for 2020 Rust toolchains.
corehack is gone! Support in core is sufficient now. At least it seems
to work...

Had to make some changes to remove the assumption that code can always
get to a waker through a thread local. That is no longer true, and was
always a bit of a hack! This should be much better.

Still on nightly because I foolishly thought the never type was near
stabilization as of last fall... but it's still unstable. But I like it.
So I'm on nightly.