exec: fix interrupt check for vector #256
ARMv7-M allows for 9 bits of interrupt number, or 512 unique vectors.
When asserting that certain internal API was not being called from
interrupt mode, we were only checking the bottom 8 bits. This means we'd
permit the API to be used within the ISR for vector #256 (hex 0x1_00).

These asserts exist to detect mistakes in the `exec` implementation, and
these mistakes don't currently exist, so this is a cleanup rather than a
bug fix.