Fix cancel behavior in mutex, queue.
There were cases in both Mutex and Queue where a wake event could be
lost, by unfortunate cancellation timing -- between the time that the
future was awoken, and the next time it got polled. This ordering of
events is unusual but _entirely possible_ and so must be addressed.

This change generalizes the solution to an optional feature on
`List::insert_and_wait` -- the caller can now provide a closure that
will be invoked on this event, and can e.g. wake another task.

Also noticed a potential bug in the Mutex implementation that could
allow a double-take, because I wasn't setting the taken bit on
contention. This is what I get for shipping code that I've used exactly
once, and in none of the examples! Time to write some tests.