mutex: improve fairness.
There was a subtle race condition between releasing the mutex in one
task and acquiring it in the next, which I have worked around by leaving
the taken flag set in that kind of "pass" scenario.

Doing this efficiently suggested a breaking change to `List::wake_one`,
which now returns a flag. Some ripple effects from that.