list: remove three panics in release.
Two panics have become debug_assertions, related to the structure of the
list -- since the change to blocking insert futures in 2020, the list
invariants that (1) Nodes reachable through &mut are not members of any
list and (2) a list being dropped is empty, are both enforced
statically.

The final panic was on double-execution of a cleanup action on a
WaitForDetach future, which ... shouldn't be possible, because it would
require a double-Drop of the future. That being said, it is possible to
write code that drops a value twice, and in this case it wouldn't
actually cause a problem if you did so, so there's no point in spending
flash space on the panic.

These together save a few hundred bytes in a typical application.