spsc: remove an overflowing add
Since we keep the indices bounded between 0 and storage.len()
(exclusive!) we know that any index + 1 is representable without
overflow. The compiler, however, hasn't been looped in on this fact, and
generates overflow checks for push and pop.

This switches the addition to explicitly wrapping for cheapest machine
code, with no change in semantics due to the invariant described above.