Switch internal mask computation to rotates for better codegen.
The `rotate_left` operation was not actually present in the standard
library when I wrote the original executor! It's really nice though.

In particular, this makes mask computation a few instructions shorter on
Cortex-M0, where arithmetic-by-a-constant is relatively expensive. Where
we previously had

    movs r0, #0x1F      ; get 32 (number of bits) mask
    ands r1, r0         ; compute r1 mod 32
    movs r0, #1         ; get a 1 bit
    lsls r1, r0         ; compute 1 << (i % 32)

we now get

    negs r2, r1         ; Compute right-rotate distance
    movs r0, #1         ; Get a 1 bit
    rors r0, r2         ; Compute 1.rotate_right(-i) == 1.rotate_left(i)

Mask computation is not _super_ frequent, but, this does save some text
(and imo makes the code clearer).

This has a side effect of making the core part of the executor no longer
32-bit specific -- it will adapt to the size of usize. Not that I have
any immediate plans for a 64-bit port (though I definitely accept
donations of devel hardware).