spsc: change lifetime bounds on reserve
The way I wrote this originally, the reserve future -- and the Permit
it resolved to -- both took a shared borrow on the Queue. This is
correct, however, it allows for a deadlock if you do something
pathological like this:

    join! {
        push_end_of_queue.reserve(),
        push_end_of_queue.reserve(),
    }

As tempting as it is to suggest _not doing that,_ by switching it to
borrowing the Push end exclusively, I can make this impossible.

I'm treating this as a minor-rev bugfix because it won't break any code
that doesn't already contain a deadlock.