Use pin-project-lite to, well, pin-project
I was dimly aware of pin-project's existence but had never really
evaluated it. I noticed pin-project-lite in my existing dependency graph
and thought I'd give it a shot.

Net effect: removed a bunch of unsafe blocks, down 20 lines of code, but
also had to remove documentation on a bunch of struct fields because it
can't handle doc attrs.

In my case the fields are private, so, this is tolerable.

Codegen is very slightly different but in ways that appear to be noise,
e.g. some things are 4 bytes larger, some are 4 bytes smaller, but all
in all things seem roughly equivalent.