list: fix soundness issue for insert_and_wait
For anyone keeping score at home, this is the third commit in five years
fixing soundness issues around List::insert operations. Lists are hard.

List::insert_and_wait and kin were eagerly attaching the node to the
list and then returning the wait future. This means the node gets
attached before the wait future gets pinned. This might not look like a
problem, since the wait future _itself_ is actually Unpin and doesn't
look like it contains internal references.

However -- pinning the wait future prevents it from being given to
core::mem::forget. If you forget the wait future instead of dropping it,
you regain &mut access to the Node, which is still linked into the list,
since the wait future's drop impl didn't have an opportunity to remove
it.

I have finally realized that a peculiar behavior of async fns -- that
they don't do anything until polled, and merely capture their arguments
-- is in fact a critical correctness property, and the fact that I had
"optimized" insert_and_wait to not do this created a soundness issue.

This commit moves the linking into the list to happen after first poll,
because poll requires pinning, and pinning prevents forget -- once
pinned, we will either keep living, or be dropped, but not forgotten.

This added a requirement that the wait future reference the lifetime of
the list, in addition to that of the node. TBH that probably should have
been the case already, but, now it is required to compile. This in turn
required rearranging how the timer list is tracked, because the old
set_timer_list / with_timer_list apparatus used universally quantified
lifetimes to prevent reference escape... which actually prevented most
useful uses of the wait future. In particular, it prevented the future
from escaping the with_timer_list closure where it could be await-ed!

This actually resolves a couple of apparently unrelated TODOs in the
process, and applications are building slightly smaller after these
changes.