Switch back to a workspace.
I switched away from a Cargo Workspace in the days before resolver="2"
was an option -- it's required to get feature resolution to be separate
per binary target in a workspace, which is in turn required because of
how the PACs abuse Cargo features.

That's been fixed for like three years, however, and a workspace
slightly reduces build times and greatly reduces target directory size.

The interaction between workspaces and .cargo/config.toml is still
frustrating -- config.toml is still the only way to set several critical
build options, but is ignored unless you CD into the relevant directory
before building, which is otherwise not required in a workspace. But,
I'll give this a shot.

Before:
- Clean build: 2m27s
- No-change rebuild: 0.428s

After:
- Clean build: 30s (!)
- rebuild: 16s :/