list: fix subtle lurking bug in ordering
The list structure has had a subtle potential bug for quite some time,
but it only became observable with the addition of wake_while.

For most of lilos's history there have been two broad classes of lists:

- Queue-style lists where nodes are not ordered with respect to one
  another (they use ordering type `()`) and nodes are removed one at a
  time.

- Timer-style lists where nodes are sorted and are removed in batches
  based on a predicate (e.g. `contents < timestamp`).

This was effectively enforced by the API, which had only two ways to
remove a node: wake_one (implemented only for ordering type `()`) and
wake_less_than. Because the batch-wake predicate was fixed at "less
than," all the nodes with a given contents value (e.g. timestamp) would
be detached during the same operation.

It turns out, this was hiding the fact that the nodes were being
inserted backwards. Nodes would be inserted into the list at the _front_
of their "equivalence group" (a run of nodes with the same contents
value). wake_less_than processes from front to back, and was thus waking
the most recently inserted node first, but this fact was not
_observable_ because all nodes with the same value would be detached
before any other code could run.

wake_one, presumably to fix a bug at some point in the past, was
removing one node from the _back_ of the list.

This meant that mixing wake_one and wake_less_than on the same list
could produce weird results, though that never happened in practice.
More importantly, it meant that wake_while -- which is capable of
detecting iteration order in various ways -- produced unfair results,
waking nodes in observably LIFO order.

This change...

- Adds a list test case that reproduces the issue by abusing wake_while.
- Fixes the list wake and insert algorithms to be consistent.