os: switch from "secret features" to non-feature cfgs
This moves basepri and native atomics detection to cfgs (namespaced by
lilos_) instead of fake features. This seems less confusing, since
normally if you see cfg(feature = "foo") in the source, you can expect
to also find a definition of that feature in Cargo.toml... which was not
the case here.