spsc: fix split lifetimes
In the current implementation, it was not possible to split the queue in
a function where the queue was passed as a mutable reference. For
example when wrapping Pusher and Popper in custom types. This failed
with an "`queue` dropped here while still borrowed, borrow might be used
here, when `queue` is dropped and runs the `Drop` code for type
`spsc::Queue`". This commit attempts to solve it by introducing an
explicit lifetime of the &mut capture when splitting, which is shorter
than the lifetime of the underlying data.

https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=17cad726a74a1be401924d2686e757ab