Add `idle` parameter to idle callback
This changes the semantics of the idle callback to be called even if
there are still runnable tasks. This allows it to be used to also poll
for events rather than being entirely interrupt driven.

It now takes an `idle` bool parameter which tells the callback whether
there are any active tasks. Using a body of the form

```
|idle| if idle { ... }
```
has the same semantics as before.